Removed 'DIN Medium.ttf' due to license issues!
I have received the following email,
hence delete the file in question.

~Sven

+++

Manja Seltmann
Paralegal Enforcement Europe

for Jutta Kurth
Legal Counsel Enforcement
Telephone XXXXXXXXXXX

Monotype GmbH
Werner-Reimers-Str. 2-4
61352 Bad Homburg

++++

Dear Mr. Göthel,

Please allow me to introduce myself as Legal Counsel Enforcement of Monotype GmbH (“Monotype”). Monotype (together with its sister company Monotype Imaging Inc.) is a leading provider of type, technology and expertise, which offers one of the world’s largest and most highly regarded typeface libraries.

We noticed that your parent directory is published on the Internet and therefore the contents thereof available to everybody. You will find the relevant link below:

http://jogamp.org/doc/siggraph2013/bof/mark/

We regularly search the Internet for websites on which our trademark and copyright protected fonts are available for free download. Your parent directory comprises a folder “fonts” in which of one of our fonts is included.

DIN®

http://www.uspto.gov/trademarks/index.jsp

Registration Number: 75642456

This font name is a registered trademark and therefore Monotype’s intellectual property. We have stated the link to the trademark search engine of the respective trademark offices.

By clicking on these links you will find the relevant registration details for the trademark, in particular that Monotype GmbH and/or another company belonging to the Monotype group of companies (together referred to as “Monotype”) is the owner of the trademarks.

For your information: Currently, some of trademarks are still registered for FontShop (e. g. DIN). However, due the acquisition of FontShop by Monotype GmbH (effective October 1, 2014) the trademarks are in the process of being transferred to Monotype GmbH. Although the transfer might still not be reflected in the registration details maintained by the trademark offices Monotype GmbH is already the owner of the trademarks

As mentioned, besides the name being a protected trademark, the fonts offered under the trademark are also software programs which benefit from copyright protection.

Although we assume that you have legitimately purchased these fonts, under the end user license agreement (“EULA”) you are not entitled to grant access to these fonts for free. Monotype does not grant any licenses which include the right to allow the download of fonts at no cost.

As explained, making available our fonts to Internet users constitutes an infringement of your contractual obligations under the EULA. We therefore ask you to immediately bar access to the infringing content by

March 12, 2015.

If you have any further queries in connection with the foregoing, please do not hesitate to contact us.

Yours sincerely,